Refactor file structure.
I did some minor abstraction of some methods such that it will be easier for
us to work async over sat.
