+++ .gitignore
@@ -1,0 +1,5 @@
+venv/
+.vscode/
+data/
+__pycache__/
+*.pyc
